[kafka streams] Exposing topology via producer
