+++ kafka-streams-quickstart/aggregator/src/main/java/org/acme/quarkus/sample/kafkastreams/streams/TopologyProducer.java
@@ -1,0 +1,73 @@
+package org.acme.quarkus.sample.kafkastreams.streams;
+
+import java.time.Instant;
+
+import javax.enterprise.context.ApplicationScoped;
+import javax.enterprise.inject.Produces;
+
+import org.acme.quarkus.sample.kafkastreams.model.Aggregation;
+import org.acme.quarkus.sample.kafkastreams.model.TemperatureMeasurement;
+import org.acme.quarkus.sample.kafkastreams.model.WeatherStation;
+import org.apache.kafka.common.serialization.Serdes;
+import org.apache.kafka.streams.StreamsBuilder;
+import org.apache.kafka.streams.Topology;
+import org.apache.kafka.streams.kstream.Consumed;
+import org.apache.kafka.streams.kstream.GlobalKTable;
+import org.apache.kafka.streams.kstream.Materialized;
+import org.apache.kafka.streams.kstream.Produced;
+import org.apache.kafka.streams.state.KeyValueBytesStoreSupplier;
+import org.apache.kafka.streams.state.Stores;
+
+import io.quarkus.kafka.client.serialization.JsonbSerde;
+
+@ApplicationScoped
+public class TopologyProducer {
+
+    static final String WEATHER_STATIONS_STORE = "weather-stations-store";
+
+    private static final String WEATHER_STATIONS_TOPIC = "weather-stations";
+    private static final String TEMPERATURE_VALUES_TOPIC = "temperature-values";
+    private static final String TEMPERATURES_AGGREGATED_TOPIC = "temperatures-aggregated";
+
+    @Produces
+    public Topology buildTopology() {
+        StreamsBuilder builder = new StreamsBuilder();
+
+        JsonbSerde<WeatherStation> weatherStationSerde = new JsonbSerde<>(WeatherStation.class);
+        JsonbSerde<Aggregation> aggregationSerde = new JsonbSerde<>(Aggregation.class);
+
+        KeyValueBytesStoreSupplier storeSupplier = Stores.persistentKeyValueStore(WEATHER_STATIONS_STORE);
+
+        GlobalKTable<Integer, WeatherStation> stations = builder.globalTable(
+                WEATHER_STATIONS_TOPIC,
+                Consumed.with(Serdes.Integer(), weatherStationSerde));
+
+        builder.stream(
+                        TEMPERATURE_VALUES_TOPIC,
+                        Consumed.with(Serdes.Integer(), Serdes.String())
+                )
+                .join(
+                        stations,
+                        (stationId, timestampAndValue) -> stationId,
+                        (timestampAndValue, station) -> {
+                            String[] parts = timestampAndValue.split(";");
+                            return new TemperatureMeasurement(station.id, station.name, Instant.parse(parts[0]), Double.valueOf(parts[1]));
+                        }
+                )
+                .groupByKey()
+                .aggregate(
+                        Aggregation::new,
+                        (stationId, value, aggregation) -> aggregation.updateFrom(value),
+                        Materialized.<Integer, Aggregation> as(storeSupplier)
+                            .withKeySerde(Serdes.Integer())
+                            .withValueSerde(aggregationSerde)
+                )
+                .toStream()
+                .to(
+                        TEMPERATURES_AGGREGATED_TOPIC,
+                        Produced.with(Serdes.Integer(), aggregationSerde)
+                );
+
+        return builder.build();
+    }
+}
