+++ kafka-streams-quickstart/aggregator/src/main/resources/application.properties
@@ -1,3 +1,14 @@
-org.acme.quarkus.sample.kafkastreams.bootstrap.servers=localhost:9092
+quarkus.kafka-streams.bootstrap-servers=localhost:9092
+quarkus.kafka-streams.application-id=temperature-aggregator
+quarkus.kafka-streams.application-server=${hostname}:8080
+quarkus.kafka-streams.topics=weather-stations,temperature-values
+
+# streams options
+kafka-streams.cache.max.bytes.buffering=10240
+kafka-streams.commit.interval.ms=1000
+kafka-streams.metadata.max.age.ms=500
+kafka-streams.auto.offset.reset=earliest
+kafka-streams.metrics.recording.level=DEBUG
+
 quarkus.log.console.enable=true
 quarkus.log.console.level=INFO
